pyproject.toml
[project]
name = "canml"
version = "0.1.3"
description = "Decode CAN BLF logs using DBC files into pandas DataFrames and export to CSV"
readme = { file = "README.MD", content-type = "text/markdown" }
authors = [
  { name = "Cosmin B. Memetea", email = "cosmin.memetea@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "cantools==39.4.4",
  "python-can==4.4.0",
  "pandas==2.2.2",
  "numpy==1.26.4",
  "tqdm>=4.0.0",
  "pyarrow>=11.0.0"
]

# Trove classifiers must be a top-level array under [project]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Topic :: Scientific/Engineering :: Information Analysis"
]

[project.urls]
Homepage      = "https://github.com/cosminmemetea/canml"
Documentation = "https://canml.readthedocs.io/"
Source        = "https://github.com/cosminmemetea/canml"
Tracker       = "https://github.com/cosminmemetea/canml/issues"

[project.optional-dependencies]
test = [
  "pytest>=7.0.0",
  "coverage",
  "codecov",
  "twine",
  "build"
]

[build-system]
requires = ["setuptools>=61.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
where  = ["."]
include = ["canml", "canml.*"]
